+++ src/error.rs
@@ -10,8 +10,8 @@ pub enum Error {
     #[error(transparent)]
     PluginInvoke(#[from] tauri::plugin::mobile::PluginInvokeError),
     #[cfg(desktop)]
-    #[error("tauri-plugin-secure-keystore does not support desktop yet. Use tauri-plugin-stronghold (or similar) for desktop until this lands.")]
-    UnsupportedPlatform,
+    #[error(transparent)]
+    Keyring(#[from] keyring::Error),
 }
 
 impl Serialize for Error {